[IMP] l10n_fr_hr_holidays: Move calendar methods to resources
Purpose
=======

Allow extended usage in other applications (here l10n_in_hr_holidays)
