+++ addons/l10n_fr_hr_holidays/models/__init__.py
@@ -3,5 +3,4 @@
 
 from . import hr_leave
 from . import res_company
-from . import resource_calendar
 from . import res_config_settings
