+++ addons/resource/models/resource_calendar.py
@@ -18,7 +18,7 @@ from odoo.exceptions import ValidationError
 from odoo.osv import expression
 from odoo.tools.float_utils import float_round
 
-from odoo.tools import date_utils, float_utils
+from odoo.tools import date_utils, ormcache
 from .utils import Intervals, float_to_time, make_aware, datetime_to_string, string_to_datetime
 
 
@@ -748,3 +748,22 @@ class ResourceCalendar(models.Model):
         for attendance in self.attendance_ids.filtered(lambda a: a.day_period != 'lunch' and ((not a.date_from or not a.date_to) or (a.date_from <= end.date() and a.date_to >= start.date()))):
             mapped_data[(attendance.week_type, attendance.dayofweek)] += attendance.hour_to - attendance.hour_from
         return max(mapped_data.values())
+
+    def _works_on_date(self, date):
+        self.ensure_one()
+
+        working_days = self._get_working_hours()
+        dayofweek = str(date.weekday())
+        if self.two_weeks_calendar:
+            weektype = str(self.env['resource.calendar.attendance'].get_week_type(date))
+            return working_days[weektype][dayofweek]
+        return working_days[False][dayofweek]
+
+    @ormcache('self.id')
+    def _get_working_hours(self):
+        self.ensure_one()
+
+        working_days = defaultdict(lambda: defaultdict(lambda: False))
+        for attendance in self.attendance_ids:
+            working_days[attendance.week_type][attendance.dayofweek] = True
+        return working_days
